lightningd/chaintopology: use plugin backend to setup topology
This adds `getchaininfo` and `getrawblockbyheight` handling lightningd-side,
and use them in setup_topology(). Then remove legacy bitcoind_getblockcount()
(we already get the count in `getchaininfo`), bitcoind_getblockchaininfo()
(it was only used in setup_topology()), and wait_for_bitcoind() (this was specific
to bitcoin-core and we assume our Bitcoin backend to be functional if the plugin
responds to `init`).