lightning/bitcoind: batch fees estimation
We are only interested in having an urgent, a normal, and a slow fee
estimations, but we still kept the bitcoin-core-like logic (modes and
targets) after having externalized the Bitcoin backend.

This makes our request for fees much simpler, and leaves some more
flexibility for a plugin to do something smart (it could still lie before
but now it's explicit, at least.) as we don't explicitly request
estimation for a specific mode and a target.
In addition we only make one request for all the three estimations.

Changelog-Changed: We now batch the requests for fee estimation to our Bitcoin backend.